split `BattleDetailsPatch`
